refactor(services): reorganize service types and streamline imports across clipboard, db, download, http, notifications, os, sound, and transcription services
