+++ apps/app/src/lib/services/db/index.ts
@@ -1,2 +1,19 @@
-export * from './dexie';
-export * from './models';
+export type { DbService } from './types';
+export { createDbServiceDexie } from './dexie';
+export {
+	TRANSFORMATION_STEP_TYPES,
+	TRANSFORMATION_STEP_TYPES_TO_LABELS,
+	generateDefaultTransformation,
+	generateDefaultTransformationStep,
+} from './models';
+export type {
+	Recording,
+	TransformationStepRun,
+	TransformationRun,
+	Transformation,
+	TransformationRunCompleted,
+	TransformationRunFailed,
+	DbServiceError,
+	InsertTransformationStep,
+	TransformationStep,
+} from './models';
