+++ apps/app/src/lib/services/db/types.ts
@@ -1,6 +1,6 @@
 import type { Result } from '@epicenterhq/result';
 import type {
-	DbServiceErrorProperties,
+	DbServiceError,
 	Recording,
 	TransformationStepRun,
 	TransformationRun,
@@ -11,30 +11,24 @@ import type {
 
 export type DbService = {
 	// Recording methods
-	getAllRecordings: () => Promise<
-		Result<Recording[], DbServiceErrorProperties>
-	>;
-	getLatestRecording: () => Promise<
-		Result<Recording | null, DbServiceErrorProperties>
-	>;
-	getTranscribingRecordingIds: () => Promise<
-		Result<string[], DbServiceErrorProperties>
-	>;
+	getAllRecordings: () => Promise<Result<Recording[], DbServiceError>>;
+	getLatestRecording: () => Promise<Result<Recording | null, DbServiceError>>;
+	getTranscribingRecordingIds: () => Promise<Result<string[], DbServiceError>>;
 	getRecordingById: (
 		id: string,
-	) => Promise<Result<Recording | null, DbServiceErrorProperties>>;
+	) => Promise<Result<Recording | null, DbServiceError>>;
 	createRecording: (
 		recording: Recording,
-	) => Promise<Result<Recording, DbServiceErrorProperties>>;
+	) => Promise<Result<Recording, DbServiceError>>;
 	updateRecording: (
 		recording: Recording,
-	) => Promise<Result<Recording, DbServiceErrorProperties>>;
+	) => Promise<Result<Recording, DbServiceError>>;
 	deleteRecording: (
 		recording: Recording,
-	) => Promise<Result<void, DbServiceErrorProperties>>;
+	) => Promise<Result<void, DbServiceError>>;
 	deleteRecordings: (
 		recordings: Recording[],
-	) => Promise<Result<void, DbServiceErrorProperties>>;
+	) => Promise<Result<void, DbServiceError>>;
 	/**
 	 * Checks and deletes expired recordings based on current settings.
 	 * This should be called:
@@ -42,63 +36,61 @@ export type DbService = {
 	 * 2. Before adding new recordings
 	 * 3. When retention settings change
 	 */
-	cleanupExpiredRecordings: () => Promise<
-		Result<void, DbServiceErrorProperties>
-	>;
+	cleanupExpiredRecordings: () => Promise<Result<void, DbServiceError>>;
 
 	// Transformation methods
 	getAllTransformations: () => Promise<
-		Result<Transformation[], DbServiceErrorProperties>
+		Result<Transformation[], DbServiceError>
 	>;
 	getTransformationById: (
 		id: string,
-	) => Promise<Result<Transformation | null, DbServiceErrorProperties>>;
+	) => Promise<Result<Transformation | null, DbServiceError>>;
 	createTransformation: (
 		transformation: Transformation,
-	) => Promise<Result<Transformation, DbServiceErrorProperties>>;
+	) => Promise<Result<Transformation, DbServiceError>>;
 	updateTransformation: (
 		transformation: Transformation,
-	) => Promise<Result<Transformation, DbServiceErrorProperties>>;
+	) => Promise<Result<Transformation, DbServiceError>>;
 	deleteTransformation: (
 		transformation: Transformation,
-	) => Promise<Result<void, DbServiceErrorProperties>>;
+	) => Promise<Result<void, DbServiceError>>;
 	deleteTransformations: (
 		transformations: Transformation[],
-	) => Promise<Result<void, DbServiceErrorProperties>>;
+	) => Promise<Result<void, DbServiceError>>;
 
 	// Transformation run methods
 	getTransformationRunById: (
 		id: string,
-	) => Promise<Result<TransformationRun | null, DbServiceErrorProperties>>;
+	) => Promise<Result<TransformationRun | null, DbServiceError>>;
 	getTransformationRunsByTransformationId: (
 		transformationId: string,
-	) => Promise<Result<TransformationRun[], DbServiceErrorProperties>>;
+	) => Promise<Result<TransformationRun[], DbServiceError>>;
 	getTransformationRunsByRecordingId: (
 		recordingId: string,
-	) => Promise<Result<TransformationRun[], DbServiceErrorProperties>>;
+	) => Promise<Result<TransformationRun[], DbServiceError>>;
 	createTransformationRun: (
 		transformationRun: Pick<
 			TransformationRun,
 			'input' | 'transformationId' | 'recordingId'
 		>,
-	) => Promise<Result<TransformationRun, DbServiceErrorProperties>>;
+	) => Promise<Result<TransformationRun, DbServiceError>>;
 	addTransformationStepRunToTransformationRun: (opts: {
 		transformationRun: TransformationRun;
 		stepId: string;
 		input: string;
-	}) => Promise<Result<TransformationStepRun, DbServiceErrorProperties>>;
+	}) => Promise<Result<TransformationStepRun, DbServiceError>>;
 	markTransformationRunAndRunStepAsFailed: (opts: {
 		transformationRun: TransformationRun;
 		stepRunId: string;
 		error: string;
-	}) => Promise<Result<TransformationRunFailed, DbServiceErrorProperties>>;
+	}) => Promise<Result<TransformationRunFailed, DbServiceError>>;
 	markTransformationRunStepAsCompleted: (opts: {
 		transformationRun: TransformationRun;
 		stepRunId: string;
 		output: string;
-	}) => Promise<Result<TransformationRun, DbServiceErrorProperties>>;
+	}) => Promise<Result<TransformationRun, DbServiceError>>;
 	markTransformationRunAsCompleted: (opts: {
 		transformationRun: TransformationRun;
 		output: string;
-	}) => Promise<Result<TransformationRunCompleted, DbServiceErrorProperties>>;
+	}) => Promise<Result<TransformationRunCompleted, DbServiceError>>;
 };
