+++ apps/app/src/lib/services/download/index.ts
@@ -1,10 +1,1 @@
-import type { Result, TaggedError } from '@epicenterhq/result';
-
-export type DownloadServiceError = TaggedError<'DownloadServiceError'>;
-
-export type DownloadService = {
-	downloadBlob: (args: {
-		name: string;
-		blob: Blob;
-	}) => Promise<Result<void, DownloadServiceError>>;
-};
+export type { DownloadService, DownloadServiceError } from './types';
