+++ apps/app/src/lib/services/notifications/index.ts
@@ -1,11 +1,1 @@
-import type { ToastAndNotifyOptions } from '$lib/toasts';
-import type { Result, TaggedError } from '@epicenterhq/result';
-
-export type NotificationServiceError = TaggedError<'NotificationServiceError'>;
-
-export type NotificationService = {
-	notify: (
-		options: ToastAndNotifyOptions,
-	) => Promise<Result<string, NotificationServiceError>>;
-	clear: (id: string) => Promise<Result<void, NotificationServiceError>>;
-};
+export type { NotificationService, NotificationServiceError } from './types';
