+++ apps/app/src/lib/services/os/index.ts
@@ -1,8 +1,1 @@
-import type { TaggedError } from '@epicenterhq/result';
-import type { OsType } from '@tauri-apps/plugin-os';
-
-export type OsServiceError = TaggedError<'OsServiceError'>;
-
-export type OsService = {
-	type: () => OsType;
-};
+export type { OsService, OsServiceError } from './types';
