+++ apps/app/src/lib/services/os/types.ts
@@ -1,0 +1,8 @@
+import type { TaggedError } from '@epicenterhq/result';
+import type { OsType } from '@tauri-apps/plugin-os';
+
+export type OsServiceError = TaggedError<'OsServiceError'>;
+
+export type OsService = {
+	type: () => OsType;
+};
