+++ apps/app/src/lib/services/sound/index.ts
@@ -1,10 +1,1 @@
-import type { Result, TaggedError } from '@epicenterhq/result';
-import type { WhisperingSoundNames } from '$lib/constants';
-
-export type PlaySoundServiceError = TaggedError<'PlaySoundServiceError'>;
-
-export type PlaySoundService = {
-	playSound: (
-		soundName: WhisperingSoundNames,
-	) => Promise<Result<void, PlaySoundServiceError>>;
-};
+export type { PlaySoundService, PlaySoundServiceError } from './types';
